+++ src/pwshBedrock/Private/Add-ModelCostEstimate.ps1
@@ -86,6 +86,8 @@ function Add-ModelCostEstimate {
             'anthropic.claude-3-5-sonnet-20241022-v2:0',
             'anthropic.claude-3-5-sonnet-20240620-v1:0',
             'anthropic.claude-3-7-sonnet-20250219-v1:0',
+            'anthropic.claude-opus-4-20250514-v1:0',
+            'anthropic.claude-sonnet-4-20250514-v1:0',
             'cohere.command-text-v14',
             'cohere.command-light-text-v14',
             'cohere.command-r-v1:0',
@@ -202,6 +204,14 @@ function Add-ModelCostEstimate {
                         $outputTokenCount = $Usage.'output_tokens'
                     }
                     'anthropic.claude-3-7-sonnet-20250219-v1:0' {
+                        $inputTokenCount = $Usage.'input_tokens'
+                        $outputTokenCount = $Usage.'output_tokens'
+                    }
+                    'anthropic.claude-opus-4-20250514-v1:0' {
+                        $inputTokenCount = $Usage.'input_tokens'
+                        $outputTokenCount = $Usage.'output_tokens'
+                    }
+                    'anthropic.claude-sonnet-4-20250514-v1:0' {
                         $inputTokenCount = $Usage.'input_tokens'
                         $outputTokenCount = $Usage.'output_tokens'
                     }