+++ src/pwshBedrock/Private/Format-AnthropicMessage.ps1
@@ -110,6 +110,8 @@ function Format-AnthropicMessage {
             'anthropic.claude-3-5-sonnet-20241022-v2:0',
             'anthropic.claude-3-5-sonnet-20240620-v1:0',
             'anthropic.claude-3-7-sonnet-20250219-v1:0',
+            'anthropic.claude-opus-4-20250514-v1:0',
+            'anthropic.claude-sonnet-4-20250514-v1:0',
             'anthropic.claude-3-opus-20240229-v1:0'
         )]
         [string]$ModelID,