+++ src/pwshBedrock/Public/Invoke-AnthropicModel.ps1
@@ -12,21 +12,21 @@
     If you are providing Tools to enable Function Calling, it is recommended that you use the ReturnFullObject parameter to capture the full response object.
     See the pwshBedrock documentation for more information on Function Calling and the Anthropic model.
 .EXAMPLE
-    Invoke-AnthropicModel -Message 'Explain zero-point energy.' -ModelID 'anthropic.claude-3-5-haiku-20241022-v1:0' -Credential $awsCredential -Region 'us-west-2'
+    Invoke-AnthropicModel -Message 'Explain zero-point energy.' -ModelID 'anthropic.claude-sonnet-4-20250514-v1:0' -Credential $awsCredential -Region 'us-west-2'
 
     Sends a text message to the on-demand Anthropic model in the specified AWS region and returns the response.
 .EXAMPLE
-    Invoke-AnthropicModel -Message 'Explain zero-point energy.' -ModelID 'anthropic.claude-3-5-haiku-20241022-v1:0' -ProfileName default -Region 'us-west-2' -ReturnFullObject
+    Invoke-AnthropicModel -Message 'Explain zero-point energy.' -ModelID 'anthropic.claude-sonnet-4-20250514-v1:0' -ProfileName default -Region 'us-west-2' -ReturnFullObject
 
     Sends a text message to the on-demand Anthropic model in the specified AWS region and returns the full response object.
 .EXAMPLE
-    Invoke-AnthropicModel -Message 'Explain zero-point energy.' -ModelID 'anthropic.claude-3-5-haiku-20241022-v1:0' -ProfileName default -Region 'us-west-2' -NoContextPersist
+    Invoke-AnthropicModel -Message 'Explain zero-point energy.' -ModelID 'anthropic.claude-sonnet-4-20250514-v1:0' -ProfileName default -Region 'us-west-2' -NoContextPersist
 
     Sends a text message to the on-demand Anthropic model in the specified AWS region without persisting the conversation context history. This is useful for one-off interactions.
 .EXAMPLE
     $invokeAnthropicModelSplat = @{
         Message    = 'What can you tell me about this picture? Is it referencing something?'
-        ModelID    = 'anthropic.claude-3-sonnet-20240229-v1:0'
+        ModelID    = 'anthropic.claude-sonnet-4-20250514-v1:0'
         MediaPath  = 'C:\images\tanagra.jpg'
         AccessKey  = 'xxxxxxxxxxxxxxxxxxxx'
         SecretKey  = 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
@@ -38,7 +38,7 @@
 .EXAMPLE
     $invokeAnthropicModelSplat = @{
         Message          = 'Give a brief synopsis to your class of students of what this picture represented a hundreds years ago.'
-        ModelID          = 'anthropic.claude-3-sonnet-20240229-v1:0'
+        ModelID          = 'anthropic.claude-sonnet-4-20250514-v1:0'
         MediaPath        = 'C:\images\tanagra.jpg'
         Temperature      = 1
         SystemPrompt     = 'You are a historian from the future who has studied the provided photo for many years.'
@@ -51,7 +51,7 @@
 .EXAMPLE
     $invokeAnthropicModelSplat = @{
         Message          = 'Can you name all of the Star Fleet captains featured in the various shows over the years?'
-        ModelID          = 'anthropic.claude-3-sonnet-20240229-v1:0'
+        ModelID          = 'anthropic.claude-sonnet-4-20250514-v1:0'
         SystemPrompt     = 'You are an expert on all things Star Trek, having studied the show for decades. You often win Star Trek Trivia contests and enjoy sharing your vast knowledge of Star Trek with others.'
         Temperature      = 1
         StopSequences    = 'Picard'
@@ -65,7 +65,7 @@
 .EXAMPLE
     $invokeAnthropicModelSplat = @{
         Message              = 'Can you name all of the Star Fleet captains featured in the various shows over the years?'
-        ModelID              = 'anthropic.claude-3-7-sonnet-20250219-v1:0'
+        ModelID              = 'anthropic.claude-opus-4-20250514-v1:0'
         SystemPrompt         = 'You are an expert on all things Star Trek, having studied the show for decades. You often win Star Trek Trivia contests and enjoy sharing your vast knowledge of Star Trek with others.'
         Thinking             = $true
         ThinkingBudgetTokens = 2000
@@ -83,7 +83,7 @@
 .EXAMPLE
     $invokeAnthropicModelSplat = @{
         Message          = 'Can you name all of the Star Fleet captains featured in the various shows over the years?'
-        ModelID          = 'anthropic.claude-3-5-sonnet-20240620-v1:0'
+        ModelID          = 'anthropic.claude-sonnet-4-20250514-v1:0'
         SystemPrompt     = 'You are an expert on all things Star Trek, having studied the show for decades. You often win Star Trek Trivia contests and enjoy sharing your vast knowledge of Star Trek with others.'
         Tools            = $starTrekTriviaFunctionTool
         ToolChoice       = 'auto'
@@ -97,7 +97,7 @@
 .EXAMPLE
     $invokeAnthropicModelSplat = @{
         ToolsResults = $standardToolResult
-        ModelID      = 'anthropic.claude-3-5-sonnet-20240620-v1:0'
+        ModelID      = 'anthropic.claude-sonnet-4-20250514-v1:0'
         Credential   = $credential
         Region       = 'us-west-2'
     }
@@ -140,14 +140,14 @@
     Recommended for advanced use cases only. You usually only need to use temperature.
 .PARAMETER Thinking
     Claude will show its reasoning process through thinking content blocks in the response.
-    This switch is only available for Claude 3.7 model.
+    This switch is available for Claude 3.7 and Claude 4 models.
     Thinking is not compatible with temperature, top_p, or top_k modifications, as well as forced tool use.
     Thinking is only viewable in the full response object, and will not be included in the message history context or the message reply.
 .PARAMETER ThinkingBudgetTokens
     Maximum number of tokens that Claude may use for its internal reasoning process.
     thinking budget tokens must always be less than the max tokens you specify in your request.
     This is a required parameter when using the Thinking switch.
-    This parameter is only available for Claude 3.7 model.
+    This parameter is available for Claude 3.7 and Claude 4 models.
     This parameter has no effect if not using the Thinking switch.
 .PARAMETER Tools
     Definitions of tools that the model may use.
@@ -204,7 +204,7 @@
 .LINK
     https://docs.aws.amazon.com/bedrock/latest/userguide/model-parameters-anthropic-claude-messages.html
 .LINK
-    https://docs.aws.amazon.com/bedrock/latest/userguide/model-parameters-anthropic-claude-37.html
+    https://docs.aws.amazon.com/bedrock/latest/userguide/claude-messages-extended-thinking.html
 .LINK
     https://docs.anthropic.com/en/docs/models-overview
 .LINK
@@ -257,6 +257,8 @@ function Invoke-AnthropicModel {
             'anthropic.claude-3-5-sonnet-20241022-v2:0',
             'anthropic.claude-3-5-sonnet-20240620-v1:0',
             'anthropic.claude-3-7-sonnet-20250219-v1:0',
+            'anthropic.claude-opus-4-20250514-v1:0',
+            'anthropic.claude-sonnet-4-20250514-v1:0',
             'anthropic.claude-3-opus-20240229-v1:0'
         )]
         [string]$ModelID,
@@ -377,10 +379,15 @@ function Invoke-AnthropicModel {
     $modelInfo = $script:anthropicModelInfo | Where-Object { $_.ModelId -eq $ModelID }
     Write-Debug -Message 'Model Info:'
     Write-Debug -Message ($modelInfo | Out-String)
-
     if ($Thinking) {
-        if ($ModelID -ne 'anthropic.claude-3-7-sonnet-20250219-v1:0') {
-            throw 'Thinking is only available for the Claude 3.7 model.'
+        # Check if the model supports Thinking (Claude 3.7 and Claude 4 models)
+        $supportedThinkingModels = @(
+            'anthropic.claude-3-7-sonnet-20250219-v1:0',
+            'anthropic.claude-opus-4-20250514-v1:0',
+            'anthropic.claude-sonnet-4-20250514-v1:0'
+        )
+        if ($ModelID -notin $supportedThinkingModels) {
+            throw 'Thinking is only available for Claude 3.7 and Claude 4 models.'
         }
         if (-not $ThinkingBudgetTokens) {
             throw 'ThinkingBudgetTokens is a required parameter when using the Thinking switch.'